Finished The tutorial Project

+++ Second_Project/controllers/contactController.js
@@ -3,17 +3,17 @@ const Contact = require("../models/contactModel")
 
 //@desc Get all Contacts 
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(
    async (req, res) => {
-      const contacts = await Contact.find()
+      const contacts = await Contact.find({user_id: req.user.id})
       res.status(200).json(contacts)
    }
 ) 
 
 //@desc Create a new Contacts 
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(
    async (req, res) => {
       console.log("The Request Body is :", req.body);
@@ -25,7 +25,8 @@ const createContact = asyncHandler(
       const contact = await Contact.create({
        name,
        email,
-       phone  
+       phone,
+       user_id: req.user.id 
       })
       res.status(201).json(contact)
    }
@@ -33,7 +34,7 @@ const createContact = asyncHandler(
 
 //@desc Get Individual Contact 
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(
    async (req, res) => {
       const contact = await Contact.findById(req.params.id)
@@ -47,13 +48,17 @@ const getContact = asyncHandler(
 
 //@desc Update Individual Contact 
 //@route PUT /api/contacts/:id
-//@access public 
+//@access private 
 const updateContact = asyncHandler(
    async (req, res) => {
       const contact = await Contact.findById(req.params.id)
       if(!contact){
          res.status(404)
          throw new Error("Contact Not Found!")
+      }
+      if(contact.user_id.toString() !== req.user.id){
+         res.status(403)
+         throw new Error("User Does Not Have Permission")
       }
       const updatedContact = await Contact.findByIdAndUpdate(
          req.params.id,
@@ -67,7 +72,7 @@ const updateContact = asyncHandler(
 
 //@desc Delete Individual Contact 
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(
    async (req, res) => {
       const contact = await Contact.findById(req.params.id)
@@ -75,6 +80,12 @@ const deleteContact = asyncHandler(
          res.status(404)
          throw new Error("Contact Not Found!")
       }
+
+      if(contact.user_id.toString() !== req.user.id){
+         res.status(403)
+         throw new Error("User Does Not Have Permission")
+      }
+
       await contact.deleteOne()
       res.status(200).json(contact)
    }
